✨ feat[#44] 허브 루트 생성 및 허브 루트 검색 리팩토링

+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/request/HubRouteUpdateRequestDto.java
@@ -12,6 +12,6 @@ public class HubRouteUpdateRequestDto {
   private UUID endHubId;
   private String startHubName;
   private String endHubName;
-  private Long deliveryTime;
-  private Long deliveryDistance;
+  private Integer deliveryTime;
+  private Double deliveryDistance;
 }
