+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteDetailResponseDto.java
@@ -14,8 +14,8 @@ public class HubRouteDetailResponseDto {
   private UUID endHubId;
   private String startHubName;
   private String endHubName;
-  private Long deliveryTime;
-  private Long deliveryDistance;
+  private Integer deliveryTime;
+  private Double deliveryDistance;
 
   public static HubRouteDetailResponseDto toResponse(HubRoute hubRoute) {
     return new HubRouteDetailResponseDto(
