+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteUpdateResponseDto.java
@@ -18,8 +18,8 @@ public class HubRouteUpdateResponseDto {
   private UUID endHubId;
   private String startHubName;
   private String endHubName;
-  private Long deliveryTime;
-  private Long deliveryDistance;
+  private Integer deliveryTime;
+  private Double deliveryDistance;
 
 
   public HubRouteUpdateResponseDto(HubRoute updateHubRoute) {
